Update using_tuples.py
Modify tuple part 2:
you can re-assign the value of the variable that is holding a tuple...This is because a variable is the "box" holding the tuple. Re-assigning the value inside the variable does not change the tuple. Like when you switch out a CD in a CD player, you are not modifying the contents of the CD, you're just modifying its location.